requirements.txt
selenium
bs4
webdriver-manager
python-dotenv
requests
psycopg2-binary